refactoring inbox
