+++ cli/src/messages/mod.rs
@@ -3,5 +3,5 @@ pub mod models;
 pub mod sync;
 
 pub use db::MessageDb;
-pub use models::{Message, MessageStatus, SyncStatus};
+pub use models::{Message, MessageStatus, MessageType, SyncStatus};
 pub use sync::MessageSync;
